feat(BE): ✨ implement workshop failure (#14)
* feat(BE): :sparkles: implement workshop failure

add workshop failure to the app
need to check when clients try to enter the workshop during failure => use a specific entity state

* refactor(lib): :recycle: update LogicalDateTime

add getters for the day, month, year, hour, minutes and seconds of the datetime

* refactor(BE): :recycle: update utils

update the DateTimeFrenchFormat utility to include LogicalDateTime as a parameter

* refactor(BE): :recycle: update workshop failure

add workshop states for better control over failure state.
add conditions for clients to start their workshop. It checks the workshop state before starting it.

* chore(project): :wrench: update .gitignore

* fix(BE): :bug: update workshop failure from type

don't send a workshop failure for the workshop types that can't be in failure.
i.e. the RELAXATION and HOME.